# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Modeva CMS Backend API Documentation",
        "title": "Modeva CMS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/google/callback": {
            "get": {
                "description": "Handles the callback from Google OAuth. Verifies the state token, exchanges the authorization code, retrieves user info, creates/updates the user in the database, issues a JWT cookie, and redirects the user back to the frontend.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth - Google OAuth"
                ],
                "summary": "Google OAuth callback",
                "responses": {
                    "307": {
                        "description": "Redirect to frontend after successful login"
                    },
                    "400": {
                        "description": "Invalid state or missing authorization code",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized or token exchange failure",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/store/categories/{id}": {
            "get": {
                "description": "Get single category with subcategories and product count",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "Get category details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/store/products/filters": {
            "get": {
                "description": "Get all available filters for products (categories, sizes, price range)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "Get available product filters",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/models.Pagination"
                },
                "rate_limit": {
                    "$ref": "#/definitions/models.RateLimiter"
                },
                "requested_entity": {
                    "type": "string"
                }
            }
        },
...
        "models.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 42
                },
                "total_pages": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
        "models.RateLimiter": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "remaining": {
                    "type": "integer"
                },
                "reset_at": {
                    "type": "string"
                },
                "reset_in_seconds": {
                    "type": "integer"
                }
            }
        },
...
    }
}